@@ -1,22 +1,66 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { HttpModule } from '@angular/http';
-import { RouterModule } from '@angular/router';
+import { BrowserModule } from "@angular/platform-browser";
+import { LOCALE_ID, NgModule } from "@angular/core";
+import { HttpClientModule } from "@angular/common/http";
+import { RouterModule } from "@angular/router";
 
+import { AppComponent } from "./app.component";
+import { HeaderComponent } from "./header/header.component";
+import { HomeComponent } from "./home/home.component";
 
-import { AppComponent } from './app.component'
-
+import { ROUTES } from "./app.routes";
+import { RestaurantsComponent } from "./restaurants/restaurants.component";
+import { RestaurantComponent } from "./restaurants/restaurant/restaurant.component";
+import { RestaurantsService } from "./restaurants/restaurants.service";
+import { RestaurantDetailComponent } from "./restaurant-detail/restaurant-detail.component";
+import { MenuComponent } from "./restaurant-detail/menu/menu.component";
+import { ShoppingCartComponent } from "./restaurant-detail/shopping-cart/shopping-cart.component";
+import { MenuItemComponent } from "./restaurant-detail/menu-item/menu-item.component";
+import { ReviewsComponent } from "./restaurant-detail/reviews/reviews.component";
+import { ShoppingCartServices } from "./restaurant-detail/shopping-cart/shopping-cart.service";
+import { OrderComponent } from "./order/order.component";
+import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { InputComponent } from './shared/input/input.component';
+import { RadioComponent } from './shared/radio/radio.component';
+import { OrderItemsComponent } from './order/order-items/order-items.component';
+import { OrderService } from "./order/order.service";
+import { DeliveryCostsComponent } from './order/delivery-costs/delivery-costs.component';
+import { OrderSummaryComponent } from './order-summary/order-summary.component';
+import { RatingComponent } from './shared/rating/rating.component';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    HeaderComponent,
+    HomeComponent,
+    RestaurantsComponent,
+    RestaurantComponent,
+    RestaurantDetailComponent,
+    MenuComponent,
+    ShoppingCartComponent,
+    MenuItemComponent,
+    ReviewsComponent,
+    OrderComponent,
+    InputComponent,
+    RadioComponent,
+    OrderItemsComponent,
+    DeliveryCostsComponent,
+    OrderSummaryComponent,
+    RatingComponent,
   ],
   imports: [
     BrowserModule,
-    HttpModule,
-    RouterModule
+    FormsModule,
+    ReactiveFormsModule,
+    HttpClientModule,
+    //configuração das rotas
+    RouterModule.forRoot(ROUTES),
   ],
-  providers: [],
-  bootstrap: [AppComponent]
+  providers: [
+    RestaurantsService,
+    ShoppingCartServices,
+    OrderService,
+    { provide: LOCALE_ID, useValue: "pt-BR" },
+  ],
+  bootstrap: [AppComponent],
 })
-export class AppModule { }
+export class AppModule {}
